requirements.txt
mwparserfromhell
pyyaml
nltk